Implement local dataset cache
Sometimes uploads to huggingface for bigger samples fails.
To avoid a lot of headaches we now save the datasets locally and
only optionally push to huggingface. Scripts using the datasets
will now first check whether the dataset split locally exists
and only if that is not the case try to load from huggingface.